tests: mutualize small_infra create function with etcs_infra
- move create fn to dedicated module
- add signaling_system param
